chore: 상수 파일 핫리로딩 방지
