+++ src/constants/agreeValues.constant.ts
@@ -1,0 +1,14 @@
+export const AGREE_VALUES = [
+  {
+    name: 'termsAgreement',
+    label: '이용 약관 동의(필수)',
+  },
+  {
+    name: 'privacyAgreement',
+    label: '개인정보 수집 및 이용 동의(필수)',
+  },
+  {
+    name: 'marketingAgreement',
+    label: '광고성 정보 수신 동의(선택)',
+  },
+] as const;
