+++ src/constants/index.tsx
@@ -1,56 +1,7 @@
-// 상수 키워드 추가 시 export 해주기
-
-// 상수 형태로 관리할 value들은 여기서 추가하면서 사용하기
-const AGREE_VALUES = [
-  {
-    name: 'termsAgreement',
-    label: '이용 약관 동의(필수)',
-  },
-  {
-    name: 'privacyAgreement',
-    label: '개인정보 수집 및 이용 동의(필수)',
-  },
-  {
-    name: 'marketingAgreement',
-    label: '광고성 정보 수신 동의(선택)',
-  },
-] as const;
-
-export const OFFSET_THRESHOLD = 30;
-export const DELTA_THRESHOLD = 5;
-
-export { AGREE_VALUES };
-
-export const BUTTON_DATA = [
-  { label: '공지 사항', path: '/mypage/notice' },
-  { label: '문의 사항', path: '/mypage/inquiry' },
-  { label: '이용 기록', path: '/mypage/useage' },
-  { label: '알림 설정', path: '/mypage/notification' },
-  {
-    label: '전화번호 인증',
-    path: 'mypage/phone/verification',
-    isVerified: false,
-  },
-];
-
-export const ORIGIN = '127.12692157601926,37.45052385846493';
-export const DESTINATION = '127.13467190126833,37.45543030528147';
-
-// 수동 매칭 시간 선택에 사용할 상수 모음
-export const DAY_PERIOD = ['오전', '오후'];
-
-export const HOURS = Array.from({ length: 12 }, (_, i) =>
-  (i + 1).toString().padStart(2, '0'),
-);
-
-export const MINUTES = Array.from({ length: 6 }, (_, i) =>
-  i.toString().padEnd(2, '0'),
-);
-
-export const GRADIENT_STYLE = {
-  maskImage:
-    'linear-gradient(to bottom, transparent, black 20%, black 80%, transparent)',
-  WebkitMaskImage:
-    'linear-gradient(to bottom, transparent, black 20%, black 80%, transparent)',
-  WebkitMaskComposite: 'destination-in', // Safari 호환
-};
+/* eslint-disable react-refresh/only-export-components */
+export * from './agreeValues.constant';
+export * from './thresholds.constant';
+export * from './buttonData.constant';
+export * from './routes.constant';
+export * from './timeOptions.constant';
+export * from './styles.constant';
