+++ src/constants/routes.constant.ts
@@ -1,0 +1,2 @@
+export const ORIGIN = '127.12692157601926,37.45052385846493';
+export const DESTINATION = '127.13467190126833,37.45543030528147';
