+++ src/constants/styles.constant.ts
@@ -1,0 +1,7 @@
+export const GRADIENT_STYLE = {
+  maskImage:
+    'linear-gradient(to bottom, transparent, black 20%, black 80%, transparent)',
+  WebkitMaskImage:
+    'linear-gradient(to bottom, transparent, black 20%, black 80%, transparent)',
+  WebkitMaskComposite: 'destination-in', // Safari 호환
+} as const;
