+++ src/constants/thresholds.constant.ts
@@ -1,0 +1,2 @@
+export const OFFSET_THRESHOLD = 30;
+export const DELTA_THRESHOLD = 5;
